+++ include/Operator/MSMJ/profiler/tuple_productivity_profiler.h
@@ -29,15 +29,15 @@ public:
 
     ~TupleProductivityProfiler() = default;
 
-    auto get_join_record_map() -> phmap::parallel_flat_hash_map<int, int>;
+    auto get_join_record_map() -> phmap::parallel_flat_hash_map<uint64_t, uint64_t>;
 
-    auto add_join_record(int stream_id, int count) -> void;
+    auto add_join_record(uint64_t stream_id, uint64_t count) -> void;
 
-    auto update_cross_join(int Di, int res) -> void;
+    auto update_cross_join(uint64_t Di, uint64_t res) -> void;
 
-    auto update_join_res(int Di, int res) -> void;
+    auto update_join_res(uint64_t Di, uint64_t res) -> void;
 
-    auto get_select_ratio(int K) -> double;
+    auto get_select_ratio(uint64_t K) -> double;
 
     auto get_requirement_recall() -> double;
 
@@ -49,13 +49,13 @@ private:
     std::mutex latch_;
 
     //到达join operator的元组数量记录
-    phmap::parallel_flat_hash_map<int, int> join_record_map_{};
+    phmap::parallel_flat_hash_map<uint64_t, uint64_t> join_record_map_{};
 
     //the join operator records both the number of cross-join result size,
-    phmap::btree_map<int, int> cross_join_map_{};
+    phmap::btree_map<uint64_t, uint64_t> cross_join_map_{};
 
     //the number of join results, using map for sorting
-    phmap::btree_map<int, int> join_result_map_{};
+    phmap::btree_map<uint64_t, uint64_t> join_result_map_{};
 
 };
